Outsource histogram objects into own module
